pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "joshua-agent"
version = "0.6.1"
description = "Autonomous gated software sprints. Define your team in YAML — dev, bug-hunter, QA. Gate agent issues GO/CAUTION/REVERT. Auto-deploy on GO, auto-revert on REVERT. Self-learning wiki across cycles. Works with Claude Code, Aider, Codex, or any LLM CLI. Shall we play a game?"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Jorge Vazquez", email = "jorgevazquez-vagojo@users.noreply.github.com"},
]
keywords = ["ai", "agents", "autonomous", "multi-agent", "sprint", "llm", "claude", "codex", "aider", "skills", "quality-gate"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "requests>=2.31",
    "pydantic>=2.0",
]

[project.optional-dependencies]
redis = ["redis>=5.0"]
server = ["fastapi>=0.115", "uvicorn>=0.32"]
all = ["redis>=5.0", "fastapi>=0.115", "uvicorn>=0.32"]
dev = ["pytest>=8.0", "ruff>=0.4", "fastapi>=0.115", "httpx>=0.27", "uvicorn>=0.32"]

[project.scripts]
joshua = "joshua.cli:main"

[project.urls]
Homepage = "https://github.com/jorgevazquez-vagojo/joshua-agent"
Documentation = "https://github.com/jorgevazquez-vagojo/joshua-agent#readme"
Repository = "https://github.com/jorgevazquez-vagojo/joshua-agent"
Issues = "https://github.com/jorgevazquez-vagojo/joshua-agent/issues"

[tool.setuptools.packages.find]
include = ["joshua*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
